+++ app/models/digitization/map.rb
@@ -1,4 +1,6 @@
-class Digitization::Map < ApplicationRecord
+class Digitization::Map < JupiterCore::Depositable
+
+  belongs_to :owner, class_name: 'User'
 
   validates :peel_map_id, uniqueness: true