+++ db/migrate/20210427212402_add_depositable_columns_to_digitization_image.rb
@@ -1,0 +1,16 @@
+class AddDepositableColumnsToDigitizationImage < ActiveRecord::Migration[6.0]
+  def change
+    change_table :digitization_images do |t|
+      t.datetime :date_ingested, null: false
+      t.datetime :record_created_at
+      t.string :visibility
+      t.references :owner, null: false, index: true, foreign_key: {to_table: :users, column: :id}
+      t.string :title, null: false
+    end
+
+    add_rdf_table_annotations for_table: :digitization_images do |t|
+      t.date_ingested has_predicate: RDF::Vocab::EBUCore.dateIngested
+      t.record_created_at has_predicate: ::TERMS[:ual].record_created_in_jupiter
+    end
+  end
+end