+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_04_22_200517) do
+ActiveRecord::Schema.define(version: 2021_04_27_214238) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -110,6 +110,11 @@ ActiveRecord::Schema.define(version: 2021_04_22_200517) do
     t.string "rights"
     t.string "topical_subjects", array: true
     t.string "volume_label"
+    t.datetime "date_ingested", null: false
+    t.datetime "record_created_at"
+    t.string "visibility"
+    t.bigint "owner_id", null: false
+    t.index ["owner_id"], name: "index_digitization_books_on_owner_id"
     t.index ["peel_id", "run", "part_number"], name: "unique_peel_book", unique: true
   end
 
@@ -117,6 +122,12 @@ ActiveRecord::Schema.define(version: 2021_04_22_200517) do
     t.string "peel_image_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.datetime "date_ingested", null: false
+    t.datetime "record_created_at"
+    t.string "visibility"
+    t.bigint "owner_id", null: false
+    t.string "title", null: false
+    t.index ["owner_id"], name: "index_digitization_images_on_owner_id"
     t.index ["peel_image_id"], name: "unique_peel_image", unique: true
   end
 
@@ -124,6 +135,12 @@ ActiveRecord::Schema.define(version: 2021_04_22_200517) do
     t.string "peel_map_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.datetime "date_ingested", null: false
+    t.datetime "record_created_at"
+    t.string "visibility"
+    t.bigint "owner_id", null: false
+    t.string "title", null: false
+    t.index ["owner_id"], name: "index_digitization_maps_on_owner_id"
     t.index ["peel_map_id"], name: "unique_peel_map", unique: true
   end
 
@@ -134,6 +151,12 @@ ActiveRecord::Schema.define(version: 2021_04_22_200517) do
     t.string "day"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.datetime "date_ingested", null: false
+    t.datetime "record_created_at"
+    t.string "visibility"
+    t.bigint "owner_id", null: false
+    t.string "title", null: false
+    t.index ["owner_id"], name: "index_digitization_newspapers_on_owner_id"
     t.index ["publication_code", "year", "month", "day"], name: "unique_peel_newspaper", unique: true
   end
 
@@ -401,6 +424,10 @@ ActiveRecord::Schema.define(version: 2021_04_22_200517) do
   add_foreign_key "announcements", "users"
   add_foreign_key "collections", "users", column: "owner_id"
   add_foreign_key "communities", "users", column: "owner_id"
+  add_foreign_key "digitization_books", "users", column: "owner_id"
+  add_foreign_key "digitization_images", "users", column: "owner_id"
+  add_foreign_key "digitization_maps", "users", column: "owner_id"
+  add_foreign_key "digitization_newspapers", "users", column: "owner_id"
   add_foreign_key "draft_items", "users"
   add_foreign_key "draft_theses", "institutions"
   add_foreign_key "draft_theses", "languages"