+++ CHANGELOG.md
@@ -58,6 +58,10 @@ Further discussion of the context can be found at [#2119](https://github.com/ual
 – Rails upgraded to 6.0.3.6 to resolve certain issues with community dependencies
 - Moved `visibility` vocabulary into a `jupiter_core` namespace
 - Move `doi_url` to `Doiable` class
+- Make `Digitization::Book` `Depositable`
+- Make `Digitization::Newspaper` `Depositable`
+- Make `Digitization::Image` `Depositable`
+- Make `Digitization::Map` `Depositable`
 
 ### Fixed
 - bump rubocop and fix cop violations [PR#2072](https://github.com/ualbertalib/jupiter/pull/2072)